Add stash WIP password-confirmation package
