+++ app/javascript/packages/password-confirmation/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@18f/identity-password-confirmation",
+  "private": true,
+  "version": "1.0.0",
+  "peerDependencies": {
+    "react": "^17.0.2"
+  },
+  "peerDependenciesMeta": {
+    "react": {
+      "optional": true
+    }
+  }
+}
